requirements.txt
django
psycopg[binary]
dotenv
assemblyai
openai
django-allauth
requests
pyjwt
cryptography
django-redis
django-debug-toolbar
gunicorn
yt-dlp
whitenoise
pytubefix
fastapi.responses